requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bitarray==2.9.2
blinker==1.7.0
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colbert-ai==0.2.19
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
datasets==2.17.1
debugpy==1.8.0
decorator==5.1.1
dill==0.3.8
distro==1.9.0
editdistance==0.8.1
elastic-transport==8.12.0
elasticsearch==8.12.1
entrypoints==0.4
eval_type_backport==0.2.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
Flask==3.0.2
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2023.10.0
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.42
h11==0.14.0
httpcore==1.0.3
httpx
huggingface-hub==0.23.4
idna==3.6
igraph==0.11.4
importlib-metadata==7.0.1
importlib_resources==6.1.2
ipdb==0.13.13
ipykernel==6.28.0
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.5.8
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.4
openai
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pydantic
pydantic_core==2.18.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
rank-bm25==0.2.2
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.6.0
sentencepiece==0.2.0
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
tabulate==0.9.0
texttable==1.7.0
thefuzz==0.22.1
threadpoolctl==3.3.0
tiktoken==0.6.0
together==1.1.4
tokenizers==0.15.2
tomli==2.0.1
torch
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.37.2
typer==0.12.3
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.9.0
Unidecode==1.3.8
urllib3==2.2.1
wcwidth==0.2.5
Werkzeug==3.0.1
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.1
langchain
langchain-openai
langchain-together
langchain_community
llama-cpp-python
pytrec_eval
gritlm
ftfy
easydict
fastapi
umap-learn
numpy==1.26.4
pandas
evaluate==0.4.4
optuna==4.4.0
qdrant-client
loguru
spacy